add email authentication  based on credentials file
Instead of using an email password stored on secrets.ps1 in clear text , use a credential file that provides a basic level of encryption.

While it would be relatively simple to retrieve the original password, it wouldn't be as simple as opening secrets.ps1 and that could be enough on small deployments in which you don't want to leave the email password so easily available to the end users.

Signed-off-by: Manuel Fombuena <fombuena@outlook.com>